From alexbelgium/hassio-addons

+++ battybirdnet-pi/rootfs/etc/cont-init.d/01-structure.sh
@@ -12,7 +12,7 @@ bashio::log.info "Ensuring the file structure is correct :"
 # Define structure
 echo "... creating default files"
 touch /config/include_species_list.txt # Should be null
-for files in apprise.txt exclude_species_list.txt IdentifiedSoFar.txt disk_check_exclude.txt confirmed_species_list.txt blacklisted_images.txt; do
+for files in apprise.txt exclude_species_list.txt IdentifiedSoFar.txt disk_check_exclude.txt confirmed_species_list.txt blacklisted_images.txt whitelist_species_list.txt; do
     if [ ! -f /config/"$files" ]; then
         echo "" > /config/"$files"
     fi
@@ -42,8 +42,12 @@ if df -T /tmp | grep -q "tmpfs"; then
     echo "... tmpfs detected, using it for StreamData and Processed to reduce disk wear"
     mkdir -p /tmp/StreamData
     mkdir -p /tmp/Processed
-    rm -r "$HOME"/BirdSongs/StreamData
-    rm -r "$HOME"/BirdSongs/Processed
+    if [ -d "$HOME"/BirdSongs/StreamData ]; then
+      rm -r "$HOME"/BirdSongs/StreamData
+    fi
+    if [ -d "$HOME"/BirdSongs/Processed ]; then
+      rm -r "$HOME"/BirdSongs/Processed
+    fi
     sudo -u pi ln -fs /tmp/StreamData "$HOME"/BirdSongs/StreamData
     sudo -u pi ln -fs /tmp/Processed "$HOME"/BirdSongs/Processed
 fi
@@ -54,11 +58,11 @@ chown -R pi:pi /config /etc/birdnet "$BIRDSONGS_FOLDER" /tmp
 chmod -R 755 /config /config /etc/birdnet "$BIRDSONGS_FOLDER" /tmp
 
 # Save default birdnet.conf to perform sanity check
-cp "$HOME"/BirdNET-Pi/birdnet.conf "$HOME"/BirdNET-Pi/birdnet.bak
+cp "$HOME"/BattyBirdNET-Analyzer/birdnet.conf "$HOME"/BattyBirdNET-Analyzer/birdnet.bak
 
 # Symlink files
 echo "... creating symlink"
-for files in "$HOME/BirdNET-Pi/birdnet.conf" "$HOME/BirdNET-Pi/blacklisted_images.txt" "$HOME/BirdNET-Pi/scripts/birds.db" "$HOME/BirdNET-Pi/BirdDB.txt" "$HOME/BirdNET-Pi/scripts/disk_check_exclude.txt" "$HOME/BirdNET-Pi/apprise.txt" "$HOME/BirdNET-Pi/exclude_species_list.txt" "$HOME/BirdNET-Pi/include_species_list.txt" "$HOME/BirdNET-Pi/IdentifiedSoFar.txt" "$HOME/BirdNET-Pi/confirmed_species_list.txt"; do
+for files in "$HOME/BattyBirdNET-Analyzer/birdnet.conf" "$HOME/BattyBirdNET-Analyzer/scripts/whitelist_species_list.txt" "$HOME/BattyBirdNET-Analyzer/blacklisted_images.txt" "$HOME/BattyBirdNET-Analyzer/scripts/birds.db" "$HOME/BattyBirdNET-Analyzer/BirdDB.txt" "$HOME/BattyBirdNET-Analyzer/scripts/disk_check_exclude.txt" "$HOME/BattyBirdNET-Analyzer/apprise.txt" "$HOME/BattyBirdNET-Analyzer/exclude_species_list.txt" "$HOME/BattyBirdNET-Analyzer/include_species_list.txt" "$HOME/BattyBirdNET-Analyzer/IdentifiedSoFar.txt" "$HOME/BattyBirdNET-Analyzer/scripts/confirmed_species_list.txt"; do
     filename="${files##*/}"
     if [ ! -f /config/"$filename" ]; then
         if [ -f "$files" ]; then
@@ -68,7 +72,8 @@ for files in "$HOME/BirdNET-Pi/birdnet.conf" "$HOME/BirdNET-Pi/blacklisted_image
         fi
     fi
     if [ -e "$files" ]; then rm "$files"; fi
-    sudo -u pi ln -fs /config/"$filename" "$files" || bashio::log.fatal "Symlink creation failed for $filename"
+    sudo -u pi ln -fs /config/"$filename" "$HOME/BattyBirdNET-Analyzer/$filename" || bashio::log.fatal "Symlink creation failed for $filename"
+    sudo -u pi ln -fs /config/"$filename" "$HOME/BattyBirdNET-Analyzer/scripts/$filename" || bashio::log.fatal "Symlink creation failed for $filename"
     sudo -u pi ln -fs /config/"$filename" /etc/birdnet/"$filename" || bashio::log.fatal "Symlink creation failed for $filename"
 done
